+++ clearwater-socket-factory/Makefile
@@ -1,11 +1,11 @@
 .PHONY: all
 all: clearwater_socket_factory test_client
 
-clearwater_socket_factory: clearwater_socket_factory.c
-	gcc -o clearwater_socket_factory clearwater_socket_factory.c
+clearwater_socket_factory: clearwater_socket_factory.cpp
+	g++ -o clearwater_socket_factory clearwater_socket_factory.cpp
 
 test_client: test_client.c
-	gcc -o test_client test_client.c
+	g++ -o test_client test_client.c
 
 .PHONY: clean
 clean:
