+++ debian/clearwater-socket-factory.upstart
@@ -1,7 +1,7 @@
 # clearwater-socket-factory
 #
-# This process runs in the default namespace, and enables processes running in
-# different namespaces to obtain sockets in the default namespace.
+# This process runs as root in the default namespace, and enables processes
+# running in different namespaces to obtain sockets in other namespaces.
 
 description "Clearwater socket factory server"
 
@@ -10,7 +10,35 @@ stop on runlevel [!2345]
 
 respawn
 
+MGMT_WHITELIST_DIR=/etc/clearwater-socket-factory/management.d/*
+SIGNALING_WHITELIST_DIR=/etc/clearwater-socket-factory/signaling.d/*
+
+get_daemon_args()
+{
+  . /etc/clearwater/config
+
+  if [ ! -z $signaling_namespace ]
+  then
+    signaling_namespace="--signaling_ns=$signaling_namespace"
+  fi
+
+  # The whitelists are comma separated lists of hosts. They're built up from
+  # each line of each file in the correct directory. We also remove duplicates.
+  if [ -d $MGMT_WHITELIST_DIR ]; then
+    mgmt_allowed_hosts="--mgmt_allowed_hosts=$(echo $(cat $MGMT_WHITELIST_DIR | sort | uniq) | sed -e 's/ /,/g')"
+  fi
+
+  if [ -d $SIGNALING_WHITELIST_DIR ]; then
+    signaling_allowed_hosts="--signaling_allowed_hosts=$(echo $(cat $SIGNALING_WHITELIST_DIR | sort | uniq) | sed -e 's/ /,/g')"
+  fi
+
+  DAEMON_ARGS="$signaling_namespace
+               $mgmt_allowed_hosts
+               $signaling_allowed_hosts"
+}
+
 script
   . /etc/clearwater/config
-  /usr/share/clearwater/bin/clearwater_socket_factory "$sas_server"
+  get_daemon_args
+  /usr/share/clearwater/bin/clearwater_socket_factory $DAEMON_ARGS
 end script
